+++ versions/management/commands/import_versions.py
@@ -1,130 +1,99 @@
 import djclick as click
 
 from django.core.management import call_command
+from fastcore.xtras import obj2dict
 
-from libraries.github import GithubAPIClient, GithubDataParser
-from libraries.models import Library, LibraryVersion
+from core.githubhelper import GithubAPIClient
 from versions.models import Version
+from versions.tasks import get_release_date_for_version
+
+# Minimum Boost version to import
+MIN_BOOST_VERSION = "1.10.3"
+
+# Skip beta releases, release candidates, and pre-1.0 versions
+EXCLUSIONS = ["beta", "-rc"]
+
+# Base url to generate the GitHub release URL
+BASE_GITHUB_URL = "https://github.com/boostorg/boost/releases/tag/"
 
 
 @click.command()
+@click.option("--verbose", is_flag=True, help="Enable verbose output.")
 @click.option("--delete-versions", is_flag=True, help="Delete all existing versions")
-@click.option(
-    "--skip-existing-versions",
-    is_flag=True,
-    help="Skip versions that already exist in our database",
-)
-@click.option(
-    "--delete-library-versions",
-    is_flag=True,
-    help="Delete all existing library versions",
-)
-@click.option(
-    "--create-recent-library-versions",
-    is_flag=True,
-    help=(
-        "Create library-versions for the most recent Boost version and each active "
-        "Boost library"
-    ),
-)
 @click.option("--token", is_flag=False, help="Github API token")
 def command(
+    verbose,
     delete_versions,
-    skip_existing_versions,
-    delete_library_versions,
-    create_recent_library_versions,
     token,
 ):
     """Imports Boost release information from Github and updates the local database.
 
     The function retrieves Boost tags from the main Github repo, excluding beta releases
-    and release candidates. For each tag, it fetches the associated data based on
-    whether it's a full release (data in the tag) or not (data in the commit).
+    and release candidates.
 
     It then creates or updates a Version instance in the local database for each tag.
-    Depending on the options provided, it can also delete existing versions and library
-    versions, and create new library versions for the most recent Boost version.
 
     Args:
+        verbose (bool): Enable verbose output (show logging statements)
         delete_versions (bool): If True, deletes all existing Version instances before
             importing.
-        skip-existing-versions (bool): If True, skips versions that already exist in
-            the database.
-        delete_library_versions (bool): If True, deletes all existing LibraryVersion
-            instances before importing.
-        create_recent_library_versions (bool): If True, creates a LibraryVersion for
-            each active Boost library and the most recent Boost version.
         token (str): Github API token, if you need to use something other than the
             setting.
     """
-    # Delete Versions and LibraryVersions based on options
     if delete_versions:
         Version.objects.all().delete()
         click.echo("Deleted all existing versions.")
 
-    if delete_library_versions:
-        LibraryVersion.objects.all().delete()
-        click.echo("Deleted all existing library versions.")
-
     # Get all Boost tags from Github
     client = GithubAPIClient(token=token)
     tags = client.get_tags()
+
     for tag in tags:
         name = tag["name"]
+        if verbose:
+            click.secho(f"Importing {name}...", fg="yellow")
 
-        # If we already have this version, skip importing it
-        if skip_existing_versions and Version.objects.filter(name=name).exists():
-            click.echo(f"Skipping {name}, already exists in database")
+        if skip_tag(name):
             continue
 
-        # Skip beta releases, release candidates, and pre-1.0 versions
-        if any(
-            ["beta" in name.lower(), "-rc" in name.lower(), "boost-0" in name.lower()]
-        ):
-            click.echo(f"Skipping {name}, not a full release")
-            continue
+        # Save the Version object
+        version, _ = Version.objects.update_or_create(
+            name=name,
+            defaults={"github_url": f"{BASE_GITHUB_URL}/{name}", "data": obj2dict(tag)},
+        )
 
-        # Get the metadata about the release from Github
+        # Load the release date if needed
+        if not version.release_date:
+            try:
+                get_release_date_for_version.delay(
+                    version.id, tag["commit"]["sha"], token=token
+                )
+            except Exception as e:
+                click.secho(f"Failed to load release date for {name}: {e}", fg="red")
 
-        version_data = None
-        parser = GithubDataParser()
-
-        # Try to get the metadata about the release from the tag
-        tag_data = client.get_release_by_tag(name)
-        if tag_data:
-            # This is a tag and a release, so the metadata is in the tag itself and
-            # we can parse the data we already have
-            version_data = parser.parse_tag(tag_data)
-        else:
-            # This is a tag, but not a release, so the metadata is in the commit
-            commit_data = client.get_commit_by_sha(commit_sha=tag["commit"]["sha"])
-            version_data = parser.parse_commit(commit_data)
-
-        if not version_data:
-            click.echo(f"Skipping {name}, no version data found")
-            continue
-
-        version, _ = Version.objects.update_or_create(name=name, defaults=version_data)
-        click.echo(f"Saved version {version.name}. Created: {_}")
-
+        # Load the release downloads
         add_release_downloads(version)
 
-    if create_recent_library_versions:
-        # Associate existing Libraries with the most recent LibraryVersion
-        version = Version.objects.most_recent()
-        for library in Library.objects.all():
-            library_version, _ = LibraryVersion.objects.get_or_create(
-                library=library, version=version
-            )
-            click.echo(f"Saved library version {library_version}. Created: {_}")
+        click.secho(f"Saved version {version.name}. Created: {_}", fg="green")
 
 
 def add_release_downloads(version):
     version_num = version.name.replace("boost-", "")
     if version_num < "1.63.0":
-        print(
-            "Cannot get release downloads from Artifactory for versions before 1.63.0"
-        )
         return
 
     call_command("import_artifactory_release_data", release=version_num)
+
+
+def skip_tag(name):
+    """Returns True if the given tag should be skipped."""
+    # If this version falls in our exclusion list, skip it
+    if any(pattern in name.lower() for pattern in EXCLUSIONS):
+        return True
+
+    # If this version is too old, skip it
+    version_num = name.replace("boost-", "")
+    if version_num < MIN_BOOST_VERSION:
+        return True
+
+    return False
