+++ versions/tasks.py
@@ -1,0 +1,53 @@
+import structlog
+
+from config.celery import app
+from django.conf import settings
+from core.githubhelper import GithubAPIClient, GithubDataParser
+from versions.models import Version
+
+
+logger = structlog.getLogger(__name__)
+
+
+@app.task
+def get_release_date_for_version(version_pk, commit_sha, token=None):
+    """
+    Gets and stores the release date for a Boost version using the given commit SHA.
+
+    :param version_pk: The primary key of the version to get the release date for.
+    :param commit_sha: The SHA of the commit to get the release date for.
+    """
+    try:
+        version = Version.objects.get(pk=version_pk)
+    except Version.DoesNotExist:
+        logger.error(
+            "get_release_date_for_version_no_version_found", version_pk=version_pk
+        )
+        return
+
+    if not token:
+        token = settings.GITHUB_TOKEN
+
+    parser = GithubDataParser()
+    client = GithubAPIClient(token=token)
+
+    try:
+        commit = client.get_commit_by_sha(commit_sha=commit_sha)
+    except Exception as e:
+        logger.error(
+            "get_release_date_for_version_failed",
+            version_pk=version_pk,
+            commit_sha=commit_sha,
+            e=str(e),
+        )
+        return
+
+    commit_data = parser.parse_commit(commit)
+    release_date = commit_data.get("release_date")
+
+    if release_date:
+        version.release_date = release_date
+        version.save()
+        logger.info("get_release_date_for_version_success", version_pk=version_pk)
+    else:
+        logger.error("get_release_date_for_version_error", version_pk=version_pk)
